+++ src/main/java/cz/afrosoft/whattoeat/diet/list/logic/service/MealUpdateObject.java
@@ -1,8 +1,8 @@
 package cz.afrosoft.whattoeat.diet.list.logic.service;
 
-import java.util.Optional;
-
 import cz.afrosoft.whattoeat.cookbook.recipe.logic.model.RecipeRef;
+
+import java.util.Optional;
 
 /**
  * @author tomas.rejent
@@ -11,11 +11,11 @@ public interface MealUpdateObject {
 
     Optional<Integer> getId();
 
-    Optional<Integer> getServings();
+    Optional<Float> getServings();
 
     Optional<RecipeRef> getRecipe();
 
-    MealUpdateObject setServings(int servings);
+    MealUpdateObject setServings(float servings);
 
     MealUpdateObject setRecipe(RecipeRef recipeRef);
 }
